Move encoding, is_html, text funtcion from runner/http_response to http/content_decoding.
