+++ packages/hurl/src/runner/response.rs
@@ -157,7 +157,7 @@ fn eval_implicit_body_asserts(
                         start: spec_body.space0.source_info.end.clone(),
                         end: spec_body.space0.source_info.end.clone(),
                     },
-                    inner: e,
+                    inner: RunnerError::from(e),
                     assert: true,
                 }),
             };
@@ -176,7 +176,7 @@ fn eval_implicit_body_asserts(
                         start: spec_body.space0.source_info.end.clone(),
                         end: spec_body.space0.source_info.end.clone(),
                     },
-                    inner: e,
+                    inner: RunnerError::from(e),
                     assert: true,
                 }),
             };
@@ -198,7 +198,7 @@ fn eval_implicit_body_asserts(
                         start: spec_body.space0.source_info.end.clone(),
                         end: spec_body.space0.source_info.end.clone(),
                     },
-                    inner: e,
+                    inner: RunnerError::from(e),
                     assert: true,
                 }),
             };
